financial_wellness/requirements.txt
asgiref==3.3.1
Django==3.1.6
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-js-asset==1.2.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
gunicorn==19.9.0
psycopg2-binary>=2.8